backend/requirements.txt
# Core dependencies
fastapi
uvicorn[standard]
click
pydantic
python-multipart
pymilvus
sentence-transformers
transformers
torch
spacy
unstructured[local-inference]
langdetect
crewai
ollama
pyyaml
python-dotenv
minio
tqdm
tiktoken

# Optional LLM providers
anthropic
openai

# Development dependencies
pytest
pytest-asyncio
httpx

# Authentication (Phase 1)
bcrypt
PyJWT
SQLAlchemy
Alembic # For database migrations (optional but good practice)

# Add Milvus client
milvus==2.3.3

# Add Postgres driver
psycopg[binary]

# Add Celery and Redis for background tasks
celery>=5.0
redis>=4.0

# Add others as needed
requests

# For PDF parsing
pypdf

# REMOVED jsPDF - it's a frontend library
# For generating slides/PDF
# jsPDF

# For Langchain Search Integration (if used, ensure versions match)
# langchain==...
# langchain_community==...
# langchain_openai==...
# langchain_milvus==...

pydantic-settings